implement NewClientStream
